requirements.txt
flask
flask_cors
numpy==1.16.4
opencv-python
Pillow
scipy
mtcnn
keras_vggface
keras==2.2.4
tensorflow==1.14.0
matplotlib
h5py==2.10.0
requests
configparser
werkzeug
scikit-image